Add FP8 conversion script for Hunyuan models (#351)
